Fixed UndeadArmorUtils.java (daylight sensor was off as the moon gives 15 light too and i forgot it does that...), added null checks and fixed the issue with duplicate main classes for UndeadEvents
