+++ src/main/java/xyz/roosterseatyou/mobitems/MobItems.java
@@ -5,17 +5,24 @@ import org.bukkit.plugin.java.JavaPlugin;
 import xyz.roosterseatyou.mobitems.events.TestingListener;
 import xyz.roosterseatyou.mobitems.events.farmanimalevents.sheep.FarmAnimalEvents;
 import xyz.roosterseatyou.mobitems.events.farmanimalevents.sheep.SheepListeners;
-import xyz.roosterseatyou.mobitems.itemstacks.sheep.SheepChest;
-import xyz.roosterseatyou.mobitems.itemstacks.sheep.SheepHooves;
-import xyz.roosterseatyou.mobitems.itemstacks.sheep.SheepLegs;
-import xyz.roosterseatyou.mobitems.itemstacks.sheep.SheepMask;
+import xyz.roosterseatyou.mobitems.events.undeadevents.UndeadEvents;
+import xyz.roosterseatyou.mobitems.itemstacks.farmanimal.sheep.SheepChest;
+import xyz.roosterseatyou.mobitems.itemstacks.farmanimal.sheep.SheepHooves;
+import xyz.roosterseatyou.mobitems.itemstacks.farmanimal.sheep.SheepLegs;
+import xyz.roosterseatyou.mobitems.itemstacks.farmanimal.sheep.SheepMask;
+import xyz.roosterseatyou.mobitems.itemstacks.zombie.ZombieChest;
+import xyz.roosterseatyou.mobitems.itemstacks.zombie.ZombieFeet;
+import xyz.roosterseatyou.mobitems.itemstacks.zombie.ZombieLegs;
 import xyz.roosterseatyou.mobitems.itemstacks.zombie.ZombieMask;
 
 public final class MobItems extends JavaPlugin {
 
     @Override
     public void onEnable() {
+
         items();
+        new UndeadEvents(this);
+        UndeadEvents.playerBurn();
         getServer().getPluginManager().registerEvents(new SheepListeners(), this);
         getServer().getPluginManager().registerEvents(new FarmAnimalEvents(), this);
         getServer().getPluginManager().registerEvents(new TestingListener(), this);
@@ -29,6 +36,9 @@ public final class MobItems extends JavaPlugin {
 
     public static void items(){
         ZombieMask.init();
+        ZombieChest.init();
+        ZombieLegs.init();
+        ZombieFeet.init();
         SheepMask.init();
         SheepChest.init();
         SheepLegs.init();
