+++ src/main/java/xyz/roosterseatyou/mobitems/events/undeadevents/UndeadEvents.java
@@ -3,6 +3,7 @@ import org.bukkit.entity.Entity;
 import org.bukkit.entity.Player;
 import org.bukkit.Bukkit;
 import org.bukkit.event.EventHandler;
+import org.bukkit.plugin.Plugin;
 import org.bukkit.potion.PotionEffect;
 import org.bukkit.scheduler.BukkitScheduler;
 import org.bukkit.event.Listener;
@@ -17,19 +18,25 @@ import xyz.roosterseatyou.mobitems.utils.mobarmorutils.UndeadArmorUtils;
 
 public class UndeadEvents implements Listener{
 
+    private static Plugin plugin;
+
+    public UndeadEvents(Plugin plugin){
+        UndeadEvents.plugin = plugin;
+    }
+
     public static void playerBurn(){
 
         BukkitScheduler scheduler = Bukkit.getServer().getScheduler();
-        scheduler.scheduleSyncRepeatingTask(new MobItems(), new Runnable() {
-            @Override
-            public void run(){
-                for (Player p : Bukkit.getOnlinePlayers()){
-                    if (UndeadArmorUtils.hasFullSet(p) && UndeadArmorUtils.inSunlight(p)){
-                        p.setFireTicks(200);
-                    }
+        scheduler.scheduleSyncRepeatingTask(plugin, () -> {
+            System.out.println("lol");
+            for (Player p : Bukkit.getOnlinePlayers()){
+                if (p.getInventory().getHelmet() != null &&  p.getInventory().getChestplate() != null &&
+                        p.getInventory().getLeggings() != null &&  p.getInventory().getBoots() != null &&
+                        UndeadArmorUtils.hasFullSet(p) && UndeadArmorUtils.inSunlight(p)){
+                    p.setFireTicks(200);
                 }
-
             }
+
         }, 0L, 100L);
     }
     @EventHandler
@@ -38,7 +45,8 @@ public class UndeadEvents implements Listener{
         if (ent instanceof Player){
             Player p = (Player)ent;
 
-            if (e.getDamage() > p.getHealth() && UndeadArmorUtils.hasFullSet(p)){
+            if (e.getDamage() > p.getHealth() && p.getInventory().getHelmet() != null &&  p.getInventory().getChestplate() != null &&
+                    p.getInventory().getLeggings() != null &&  p.getInventory().getBoots() != null &&UndeadArmorUtils.hasFullSet(p)){
                 p.setHealth(p.getAttribute(Attribute.GENERIC_MAX_HEALTH).getDefaultValue());
                 p.addPotionEffect(new PotionEffect(PotionEffectType.REGENERATION , 125, 2));
                 p.addPotionEffect(new PotionEffect(PotionEffectType.FIRE_RESISTANCE , 125, 1));
