+++ src/main/java/xyz/roosterseatyou/mobitems/itemstacks/farmanimal/sheep/SheepHooves.java
@@ -1,4 +1,4 @@
-package xyz.roosterseatyou.mobitems.itemstacks.sheep;
+package xyz.roosterseatyou.mobitems.itemstacks.farmanimal.sheep;
 
 import net.kyori.adventure.text.Component;
 import net.kyori.adventure.text.format.TextColor;
@@ -9,23 +9,20 @@ import org.bukkit.enchantments.Enchantment;
 import org.bukkit.inventory.ItemStack;
 import org.bukkit.inventory.meta.LeatherArmorMeta;
 
-import javax.swing.*;
 import java.util.ArrayList;
 import java.util.List;
 
-public class SheepMask {
-    public static Component ENTITY_ID = Component.text("ENTITY ID: SHEEP").color(TextColor.fromHexString("#2d3233")).decorate(TextDecoration.ITALIC);
-    public static Component CLASS_ID = Component.text("CLASS ID: FARM").color(TextColor.fromHexString("#2d3233")).decorate(TextDecoration.ITALIC);
-    public static Component NAME = Component.text("Sheep Mask");
-    public static ItemStack SHEEP_MASK;
+public class SheepHooves {
+    public static Component NAME = Component.text("Sheep Hooves");
+    public static ItemStack SHEEP_HOOVES;
 
     public static ItemStack init(){
-        sheepMask();
+        sheepHooves();
         return null;
     }
 
-    public static void sheepMask(){
-        ItemStack item = new ItemStack(Material.LEATHER_HELMET);
+    public static void sheepHooves(){
+        ItemStack item = new ItemStack(Material.LEATHER_BOOTS);
         LeatherArmorMeta meta = (LeatherArmorMeta) item.getItemMeta();
         List<Component> lore = new ArrayList<>();
         meta.displayName(NAME);
@@ -33,10 +30,10 @@ public class SheepMask {
         meta.addEnchant(Enchantment.PROTECTION_ENVIRONMENTAL, 5, true);
         lore.add(Component.text("Beep beep, I'm a Sheep..."));
         //see FarmAnimalArmorUtils
-        lore.add(ENTITY_ID);
-        lore.add(CLASS_ID);
+        lore.add(Component.text("ENTITY ID: SHEEP").color(TextColor.fromHexString("#2d3233")).decorate(TextDecoration.ITALIC));
+        lore.add(Component.text("CLASS ID: FARM").color(TextColor.fromHexString("#2d3233")).decorate(TextDecoration.ITALIC));
         meta.lore(lore);
         item.setItemMeta(meta);
-        SHEEP_MASK = item;
+        SHEEP_HOOVES = item;
     }
 }
