+++ src/main/java/xyz/roosterseatyou/mobitems/itemstacks/farmanimal/sheep/SheepMask.java
@@ -1,4 +1,4 @@
-package xyz.roosterseatyou.mobitems.itemstacks.sheep;
+package xyz.roosterseatyou.mobitems.itemstacks.farmanimal.sheep;
 
 import net.kyori.adventure.text.Component;
 import net.kyori.adventure.text.format.TextColor;
@@ -12,17 +12,19 @@ import org.bukkit.inventory.meta.LeatherArmorMeta;
 import java.util.ArrayList;
 import java.util.List;
 
-public class SheepLegs {
-    public static Component NAME = Component.text("Sheep Leggings");
-    public static ItemStack SHEEP_LEGS;
+public class SheepMask {
+    public static Component ENTITY_ID = Component.text("ENTITY ID: SHEEP").color(TextColor.fromHexString("#2d3233")).decorate(TextDecoration.ITALIC);
+    public static Component CLASS_ID = Component.text("CLASS ID: FARM").color(TextColor.fromHexString("#2d3233")).decorate(TextDecoration.ITALIC);
+    public static Component NAME = Component.text("Sheep Mask");
+    public static ItemStack SHEEP_MASK;
 
     public static ItemStack init(){
-        sheepLeggings();
+        sheepMask();
         return null;
     }
 
-    public static void sheepLeggings(){
-        ItemStack item = new ItemStack(Material.LEATHER_LEGGINGS);
+    public static void sheepMask(){
+        ItemStack item = new ItemStack(Material.LEATHER_HELMET);
         LeatherArmorMeta meta = (LeatherArmorMeta) item.getItemMeta();
         List<Component> lore = new ArrayList<>();
         meta.displayName(NAME);
@@ -30,10 +32,10 @@ public class SheepLegs {
         meta.addEnchant(Enchantment.PROTECTION_ENVIRONMENTAL, 5, true);
         lore.add(Component.text("Beep beep, I'm a Sheep..."));
         //see FarmAnimalArmorUtils
-        lore.add(Component.text("ENTITY ID: SHEEP").color(TextColor.fromHexString("#2d3233")).decorate(TextDecoration.ITALIC));
-        lore.add(Component.text("CLASS ID: FARM").color(TextColor.fromHexString("#2d3233")).decorate(TextDecoration.ITALIC));
+        lore.add(ENTITY_ID);
+        lore.add(CLASS_ID);
         meta.lore(lore);
         item.setItemMeta(meta);
-        SHEEP_LEGS = item;
+        SHEEP_MASK = item;
     }
 }
